mngr_imbue_cloud: restructure into layered sub-packages (phase 1)
Break the flat plugin into layered sub-packages governed by an import-linter
contract (high->low): plugin > cli > bake > providers > hosts > slices >
connector > config > data_types > errors > primitives.

- Move modules into plugin/ cli/ bake/ providers/ hosts/ slices/ connector/
  (cli/ and the root leaf modules config/data_types/errors/primitives unchanged).
- Isolate the slice/bare-metal subsystem in slices/; isolate the
  provider-generic pool bake in bake/ (extraction seam toward minds).
- Co-locate both provider backends in plugin/backends.py (the slice backend
  moves out of slice_provider.py; SliceVpsDockerProviderConfig stays in
  providers/slice_provider.py to keep the layer order).
- Move pluggy entry points to plugin/entrypoints.py + plugin/slice_entrypoints.py
  and update pyproject entry-point paths.
- Add the "mngr_imbue_cloud layers contract" to the root import-linter config
  plus a test_meta_ratchets.py test that enforces it.

Pure refactor: no behavior, CLI, wire-format, or schema change. apps/ imports
nothing from mngr_imbue_cloud, so this is contained to the plugin + the two
pyproject files. 253 unit/integration tests, 56 package ratchets, and the
meta-ratchets (ty, ruff, CLI docs, both import-layer contracts) all pass.
