+++ libs/mngr_imbue_cloud/imbue/mngr_imbue_cloud/cli/hosts.py
@@ -9,13 +9,13 @@ These subcommands are listing + release helpers on top of that flow.
 
 import click
 
-from imbue.mngr_imbue_cloud.auth_helper import get_active_token
 from imbue.mngr_imbue_cloud.cli._common import emit_json
 from imbue.mngr_imbue_cloud.cli._common import fail_with_json
 from imbue.mngr_imbue_cloud.cli._common import handle_imbue_cloud_errors
 from imbue.mngr_imbue_cloud.cli._common import make_connector_client
 from imbue.mngr_imbue_cloud.cli._common import make_session_store
 from imbue.mngr_imbue_cloud.cli._common import resolve_account_or_active
+from imbue.mngr_imbue_cloud.connector.auth_helper import get_active_token
 
 
 @click.group(name="hosts")
