+++ libs/mngr_imbue_cloud/imbue/mngr_imbue_cloud/connector/session_store_test.py
@@ -5,12 +5,12 @@ from pathlib import Path
 
 from pydantic import SecretStr
 
+from imbue.mngr_imbue_cloud.connector.session_store import ImbueCloudSessionStore
+from imbue.mngr_imbue_cloud.connector.session_store import _decode_jwt_exp
+from imbue.mngr_imbue_cloud.connector.session_store import make_session_from_tokens
 from imbue.mngr_imbue_cloud.data_types import AuthSession
 from imbue.mngr_imbue_cloud.primitives import ImbueCloudAccount
 from imbue.mngr_imbue_cloud.primitives import SuperTokensUserId
-from imbue.mngr_imbue_cloud.session_store import ImbueCloudSessionStore
-from imbue.mngr_imbue_cloud.session_store import _decode_jwt_exp
-from imbue.mngr_imbue_cloud.session_store import make_session_from_tokens
 
 
 def _make_session(
