+++ libs/mngr_imbue_cloud/imbue/mngr_imbue_cloud/slices/bare_metal.py
