+++ libs/mngr_imbue_cloud/imbue/mngr_imbue_cloud/slices/bare_metal_prep_test.py
@@ -1,4 +1,4 @@
-from imbue.mngr_imbue_cloud.bare_metal_prep import build_box_prep_script
+from imbue.mngr_imbue_cloud.slices.bare_metal_prep import build_box_prep_script
 
 _POOL_PUB = "ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAITESTpoolkey mngr-pool"
 _IMAGE_URL = (
