+++ libs/mngr_imbue_cloud/imbue/mngr_imbue_cloud/slices/lima_slice.py
